+++ NEWS.md
@@ -1,3 +1,12 @@
+# BLN 0.2.0 2024-12-23 
+
+## Added
+* SOMERS calculation of emissions in peat soils, function `bln_clim_somers` and associated test function
+* internal package table for SOMERS meta-model `bln_somers`
+
+## Updated
+* `ppr_tables` in dev to prepare internal package table `bln_somers`
+
 # BLN 0.1.0 2024-11-05
 First version of the R package to assess soil quality using BLN framework